+++ Sources/KeyboardKit/Input/InputSet.swift
@@ -1,5 +1,5 @@
 //
-//  KeyboardInputSet.swift
+//  InputSet.swift
 //  KeyboardKit
 //
 //  Created by Daniel Saidi on 2020-07-03.
@@ -9,10 +9,11 @@
 import Foundation
 
 /**
- A keyboard input set represents the input parts of a system
- keyboard, the center lighter input keys.
+ An input set defines the input keys on a keyboard. The keys
+ can then be used to create a layout, which defines the full
+ set of keys, including the surrounding system keys.
  */
-public class KeyboardInputSet: Equatable {
+public class InputSet: Equatable {
     
     public init(rows: KeyboardInputRows) {
         self.rows = rows
@@ -20,7 +21,7 @@ public class KeyboardInputSet: Equatable {
     
     public let rows: KeyboardInputRows
     
-    public static func == (lhs: KeyboardInputSet, rhs: KeyboardInputSet) -> Bool {
+    public static func == (lhs: InputSet, rhs: InputSet) -> Bool {
         lhs.rows == rhs.rows
     }
 }
@@ -28,14 +29,14 @@ public class KeyboardInputSet: Equatable {
 /**
  This input set can be used in alphabetic keyboards.
  */
-public class AlphabeticKeyboardInputSet: KeyboardInputSet {}
+public class AlphabeticInputSet: InputSet {}
 
 /**
  This input set can used in numeric keyboards.
  */
-public class NumericKeyboardInputSet: KeyboardInputSet {}
+public class NumericInputSet: InputSet {}
 
 /**
  This input set can be used in symbolic keyboards.
  */
-public class SymbolicKeyboardInputSet: KeyboardInputSet {}
+public class SymbolicInputSet: InputSet {}