+++ Sources/KeyboardKit/Input/Providers/InternalSwedishKeyboardInputSetProvider.swift
@@ -22,26 +22,26 @@ class InternalSwedishKeyboardInputSetProvider: DeviceSpecificInputSetProvider, L
     let device: UIDevice
     let localeKey: String = KeyboardLocale.swedish.id
     
-    var alphabeticInputSet: AlphabeticKeyboardInputSet {
-        AlphabeticKeyboardInputSet(rows: [
+    var alphabeticInputSet: AlphabeticInputSet {
+        AlphabeticInputSet(rows: [
             KeyboardInputRow("qwertyuiopå"),
             KeyboardInputRow("asdfghjklöä"),
             row(phone: "zxcvbnm", pad: "zxcvbnm,.")
         ])
     }
     
-    var numericInputSet: NumericKeyboardInputSet {
+    var numericInputSet: NumericInputSet {
         let phoneCenter: [String] = "-/:;()".chars + ["kr"] + "&@”".chars
         let padCenter: [String] = "@#".chars + ["kr"] + "&*()’”+•".chars
-        return NumericKeyboardInputSet(rows: [
+        return NumericInputSet(rows: [
             row(phone: "1234567890", pad: "1234567890`"),
             KeyboardInputRow(device.isPhone ? phoneCenter : padCenter),
             row(phone: ".,?!’", pad: "%_-=/;:,.")
         ])
     }
     
-    var symbolicInputSet: SymbolicKeyboardInputSet {
-        SymbolicKeyboardInputSet(rows: [
+    var symbolicInputSet: SymbolicInputSet {
+        SymbolicInputSet(rows: [
             row(phone: "[]{}#%^*+=", pad: "1234567890´"),
             row(phone: "_\\|~<>€$£•", pad: "€$£^[]{}—˚…"),
             row(phone: ".,?!’", pad: "§|~≠\\<>!?")